Перенос кода в отдельную директорию
